Refactor server config module
